+++ CHANGELOG.md
@@ -1,2 +1,6 @@
-# Changelog
+# BaGL Change Log
 
+## 0.2.0
+
+`IndexGeometry` now takes a type parameter `Primitive` (should be of type 
+`Point`, `Line` or `Triangle`) and extends `Iterable<Primitive>`.